Improve disconnection (+ error) handling during connection/negotiation
Previously if there was a disconnection (especially noticeable with a user-initiated disconnection)
during negotiation, then any withTimeout() call would run to completion before marking the watch
as disconnected. This wasn't a great UX - so we need to check for disconnections during negotiation.

Also explicitly log when a crash happened in the top-level connect call (no change in behaviour,
just better logging).