Notifications: mark as read to dismiss, never delete
We should never use blobdb delete on notifications (the watch doesn't expect that);
instead we modify the notification with the STATE_READ flag when dismissed on the phone.
The watch knows to dismiss if currently displayed, and also to removed any actions
in the notification app.

Notification app will now show history because we never delete them.