Emulate remote timeline locally
- Fix timeline token interface in startup.js (was never returning anything)
- Populate a real timeline token if there is one
- Add a new "Emulate remote timeline" setting; if enabled, send a dummy token as a fallback
- If setting enabled, intercept requests to timeline API webservice URLs and instead insert/delete pins locally, immediately
- Also add explicit Pebble.insertTimelinePin/Pebble.deleteTimelinePin PKJS APIs
- Did all the wiring to enable inserting pins from json format into BlobDb (adding all the string names for entities to the enums etc)

I was hoping the changes in startup.js to override XHR would work on both platforms, but they don't work on iOS, so a separate change for iOS is done in the XHR manager.

Tested with my own test app which used both legacy/token/web-based methods and new API. Also tested with a few apps from the store which interact with timeline. And tested that it doesn't break normal watchface XHR interactions.